Move egress primitives into services
Egress policy, credential materialization, header mutation, and transport helpers now live under a service-owned package:\n\n    import "github.com/valon-technologies/gestalt/server/services/egress"\n\n    policy := egress.Policy{\n        AllowedHosts:  []string{"api.example.com"},\n        DefaultAction: egress.PolicyDeny,\n    }\n    if err := policy.CheckHost("api.example.com"); err != nil {\n        return err\n    }\n\n    materialized, err := egress.MaterializeCredential(token, egress.AuthStyleBearer, nil)\n    headers := egress.ApplyHeaderMutations(nil, materialized.Headers)\n\nThis is an import-path and ownership move only. CheckHost behavior, ErrEgressDenied matching, header mutation behavior, credential auth styles, and default transport cloning are unchanged.
