+++ gestaltd/internal/mcpoauth/http_client.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 	"time"
 
-	"github.com/valon-technologies/gestalt/server/internal/egress"
+	"github.com/valon-technologies/gestalt/server/services/egress"
 )
 
 func newHTTPClient(timeout time.Duration) (*http.Client, func()) {
