+++ gestaltd/services/plugins/declarative/base_test.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 
 	"github.com/valon-technologies/gestalt/server/core"
-	"github.com/valon-technologies/gestalt/server/internal/egress"
+	"github.com/valon-technologies/gestalt/server/services/egress"
 )
 
 type mockAuth struct{}
